to: replace FixedStringWriter with DynamicStringWriter
FixedStringWriter keeps causing UAF bugs and out of bounds reads/writes
as it's a poorly designed interface. In practice, any user of it *also*
supports arbitrarily sized strings and the only reason they are using
FixedStringWriter is for performance.

However, in practice for performance reasons, it's indistiguishable from
having a DynamicStringWriter + clearing it and reusing it multiple
times. So just switch to using it, it's safer and more ergonomic
