+++ Android.bp
@@ -13898,9 +13898,9 @@ filegroup {
         "src/base/cpu_info.cc",
         "src/base/crash_keys.cc",
         "src/base/ctrl_c_handler.cc",
+        "src/base/dynamic_string_writer.cc",
         "src/base/event_fd.cc",
         "src/base/file_utils.cc",
-        "src/base/fixed_string_writer.cc",
         "src/base/getopt_compat.cc",
         "src/base/intrusive_list.cc",
         "src/base/intrusive_tree.cc",
@@ -14006,7 +14006,6 @@ filegroup {
         "src/base/circular_queue_unittest.cc",
         "src/base/dynamic_string_writer_unittest.cc",
         "src/base/file_utils_unittest.cc",
-        "src/base/fixed_string_writer_unittest.cc",
         "src/base/flags_unittest.cc",
         "src/base/flat_hash_map_unittest.cc",
         "src/base/flat_set_unittest.cc",
