+++ include/perfetto/ext/base/BUILD.gn
@@ -30,7 +30,6 @@ source_set("base") {
     "endian.h",
     "event_fd.h",
     "file_utils.h",
-    "fixed_string_writer.h",
     "flags.h",
     "flat_hash_map.h",
     "flat_hash_map_v1.h",
