+++ src/base/BUILD.gn
@@ -38,9 +38,9 @@ perfetto_component("base") {
     "cpu_info.cc",
     "crash_keys.cc",
     "ctrl_c_handler.cc",
+    "dynamic_string_writer.cc",
     "event_fd.cc",
     "file_utils.cc",
-    "fixed_string_writer.cc",
     "getopt_compat.cc",
     "intrusive_list.cc",
     "intrusive_tree.cc",
@@ -215,7 +215,6 @@ perfetto_unittest_source_set("unittests") {
     "circular_queue_unittest.cc",
     "dynamic_string_writer_unittest.cc",
     "file_utils_unittest.cc",
-    "fixed_string_writer_unittest.cc",
     "flags_unittest.cc",
     "flat_hash_map_unittest.cc",
     "flat_set_unittest.cc",
