+++ src/traceconv/trace_to_systrace.cc
@@ -27,9 +27,8 @@
 
 #include "perfetto/base/build_config.h"
 #include "perfetto/base/logging.h"
-#include "perfetto/ext/base/fixed_string_writer.h"
+#include "perfetto/ext/base/dynamic_string_writer.h"
 #include "perfetto/ext/base/string_utils.h"
-#include "perfetto/ext/base/utils.h"
 #include "perfetto/trace_processor/trace_processor.h"
 #include "src/traceconv/utils.h"
 
@@ -85,7 +84,7 @@ const char kSystemTraceEventsFooter[] =
 inline void FormatProcess(uint32_t pid,
                           uint32_t ppid,
                           const base::StringView& name,
-                          base::FixedStringWriter* writer) {
+                          base::DynamicStringWriter* writer) {
   writer->AppendLiteral("root             ");
   writer->AppendInt(pid);
   writer->AppendLiteral("     ");
@@ -98,7 +97,7 @@ inline void FormatProcess(uint32_t pid,
 inline void FormatThread(uint32_t tid,
                          uint32_t tgid,
                          const base::StringView& name,
-                         base::FixedStringWriter* writer) {
+                         base::DynamicStringWriter* writer) {
   writer->AppendLiteral("root         ");
   writer->AppendInt(tgid);
   writer->AppendChar(' ');
@@ -114,24 +113,21 @@ inline void FormatThread(uint32_t tid,
 class QueryWriter {
  public:
   QueryWriter(trace_processor::TraceProcessor* tp, TraceWriter* trace_writer)
-      : tp_(tp),
-        buffer_(base::PagedMemory::Allocate(kBufferSize)),
-        global_writer_(static_cast<char*>(buffer_.Get()), kBufferSize),
-        trace_writer_(trace_writer) {}
+      : tp_(tp), trace_writer_(trace_writer) {}
 
   template <typename Callback>
   bool RunQuery(const std::string& sql, Callback callback) {
-    char buffer[2048];
+    base::DynamicStringWriter line_writer;
     auto iterator = tp_->ExecuteQuery(sql);
     for (uint32_t rows = 0; iterator.Next(); rows++) {
-      base::FixedStringWriter line_writer(buffer, base::ArraySize(buffer));
+      line_writer.Clear();
       callback(&iterator, &line_writer);
 
-      if (global_writer_.pos() + line_writer.pos() >= global_writer_.size()) {
+      if (global_writer_.pos() + line_writer.pos() >= kFlushThreshold) {
         fprintf(stderr, "Writing row %" PRIu32 "%c", rows, kProgressChar);
         auto str = global_writer_.GetStringView();
         trace_writer_->Write(str.data(), str.size());
-        global_writer_.reset();
+        global_writer_.Clear();
       }
       global_writer_.AppendStringView(line_writer.GetStringView());
     }
@@ -146,16 +142,15 @@ class QueryWriter {
     // Flush any dangling pieces in the global writer.
     auto str = global_writer_.GetStringView();
     trace_writer_->Write(str.data(), str.size());
-    global_writer_.reset();
+    global_writer_.Clear();
     return true;
   }
 
  private:
-  static constexpr uint32_t kBufferSize = 1024u * 1024u * 16u;
+  static constexpr size_t kFlushThreshold = size_t{1024} * 1024 * 16;
 
   trace_processor::TraceProcessor* tp_ = nullptr;
-  base::PagedMemory buffer_;
-  base::FixedStringWriter global_writer_;
+  base::DynamicStringWriter global_writer_;
   TraceWriter* trace_writer_;
 };
 
@@ -167,7 +162,7 @@ int ExtractRawEvents(TraceWriter* trace_writer,
 
   static const char kRawEventsCountSql[] = "select count(1) from ftrace_event";
   uint32_t raw_events = 0;
-  auto e_callback = [&raw_events](Iterator* it, base::FixedStringWriter*) {
+  auto e_callback = [&raw_events](Iterator* it, base::DynamicStringWriter*) {
     raw_events = static_cast<uint32_t>(it->Get(0).long_value);
   };
   if (!q_writer.RunQuery(kRawEventsCountSql, e_callback))
@@ -186,7 +181,7 @@ int ExtractRawEvents(TraceWriter* trace_writer,
   fflush(stderr);
 
   auto raw_callback = [wrapped_in_json](Iterator* it,
-                                        base::FixedStringWriter* writer) {
+                                        base::DynamicStringWriter* writer) {
     const char* line = it->Get(0 /* col */).string_value;
     if (wrapped_in_json) {
       for (uint32_t i = 0; line[i] != '\0'; i++) {
@@ -310,7 +305,7 @@ int ExtractSystrace(trace_processor::TraceProcessor* tp,
     // TODO(lalitm): change this query to actually use ppid when it is exposed
     // by the process table.
     static const char kPSql[] = "select pid, 0 as ppid, name from process";
-    auto p_callback = [](Iterator* it, base::FixedStringWriter* writer) {
+    auto p_callback = [](Iterator* it, base::DynamicStringWriter* writer) {
       uint32_t pid = static_cast<uint32_t>(it->Get(0 /* col */).long_value);
       uint32_t ppid = static_cast<uint32_t>(it->Get(1 /* col */).long_value);
       const auto& name_col = it->Get(2 /* col */);
@@ -328,7 +323,7 @@ int ExtractSystrace(trace_processor::TraceProcessor* tp,
     static const char kTSql[] =
         "select tid, COALESCE(upid, 0), thread.name "
         "from thread left join process using (upid)";
-    auto t_callback = [](Iterator* it, base::FixedStringWriter* writer) {
+    auto t_callback = [](Iterator* it, base::DynamicStringWriter* writer) {
       uint32_t tid = static_cast<uint32_t>(it->Get(0 /* col */).long_value);
       uint32_t tgid = static_cast<uint32_t>(it->Get(1 /* col */).long_value);
       const auto& name_col = it->Get(2 /* col */);
